Move nav bar config to the view controller
